Fix compiler warning when using pin interrupt

+++ cores/arduino/pins_arduino.h
@@ -216,7 +216,7 @@ uint32_t pinNametoDigitalPin(PinName p);
 // Convert an analog pin number Axx to a PinName PX_n
 #define analogInputToPinName(p)     (digitalPinToPinName(analogInputToDigitalPin(p)))
 // All pins could manage EXTI
-#define digitalPinToInterrupt(p)    (digitalPinIsValid(p) ? p : NOT_AN_INTERRUPT)
+#define digitalPinToInterrupt(p)    (digitalPinIsValid(p) ? (int)p : (int)NOT_AN_INTERRUPT)
 
 #define digitalPinHasI2C(p)         (pin_in_pinmap(digitalPinToPinName(p), PinMap_I2C_SDA) ||\
                                      pin_in_pinmap(digitalPinToPinName(p), PinMap_I2C_SCL))
